refactor(products): colocate product detail components
