+++ src/app/(main)/(products)/products/[category]/[id]/_components/index.tsx
@@ -1,5 +1,4 @@
 export { ProductLikeBadge } from "./ProductLikeBadge";
-export { ProductSummary } from "./ProductSummary";
 export { ProductViewTracker } from "./ProductViewTracker";
 export { ReviewsSection } from "./ReviewsSection";
 export { ProductDetailTemplate } from "./ProductDetailTemplate";
