+++ src/app/(main)/(products)/products/[category]/[id]/_containers/index.tsx
@@ -1,0 +1,1 @@
+export { ProductSummary } from "./ProductSummary";
