+++ src/ui/components/organisms/index.ts
@@ -7,10 +7,7 @@ export * from "./FormField";
 export * from "./ImageField";
 export * from "./ImagePreviewItem";
 export * from "./PersonValueCard";
-export * from "./ProductFeatures";
 export * from "./ProductGrid";
-export * from "./ProductOptions";
-export * from "./ProductSummary";
 export * from "./RadioField";
 export * from "./RatingStars";
 export * from "./SelectField";
